remove sortTable function from food events add them to food helpers, since it's not a response handler or a request
