+++ lib/event_listeners/food_events.js
@@ -1,32 +1,10 @@
 const $ = require('jquery')
 const foodRequest = require('../ajax_requests/food_requests.js')
 const mealRequest = require('../ajax_requests/meal_requests.js')
+const foodHelper = require('../helpers/food_helpers.js')
 const url = 'https://serene-sea-75169.herokuapp.com/api/v1/'
 
 let switcher = 1
-
-const sortTable = (table, switcher) => {
-  const tableRows = `tbody.${table}-table tr`
-	var rows = $(tableRows).get();
-	rows.sort(function(a, b) {
-		const A = getVal(a);
-		const B = getVal(b);
-		if(A < B) {return -1*switcher;}
-		if(A > B) {return 1*switcher;}
-  return 0;
-  });
-  function getVal(element){
-		var v = $(element).children('td').eq(1).text().toUpperCase();
-		if($.isNumeric(v)){
- 			v = parseInt(v,10);
-		}
-		return v;
-}
-  $.each(rows, function(index, row) {
-     $(`tbody.${table}-table`).append(row);
-  });
-}
-
 
 const traverseFoodInMeals = (id, method) => {
   return fetch(`${url}meals`, {method: 'GET'})
@@ -116,7 +94,7 @@ $(document).ready(function(){
   $('th.calories-sort').on('click', function(){
     switcher *= -1
     const table = event.target.parentElement.parentElement.parentElement.id
-    sortTable(table, switcher)
+    foodHelper.sortTable(table, switcher)
  })
